# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products API",
        "description": "API para gestión de productos, órdenes y usuarios con JWT Auth.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuario y obtener token JWT",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "ana@correo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secreta123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar productos del usuario autenticado",
                "operationId": "ae5ebe0a24b2789df96322f874352c48",
                "responses": {
                    "200": {
                        "description": "Catálogo exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crea un nuevo producto",
                "operationId": "cd35318bda7d2573339631255f39f0cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Iphone 18 Ultra"
                                    },
                                    "price": {
                                        "type": "numeric",
                                        "example": 119.95
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "Error de autenticación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "required": [
                    "name",
                    "price",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Iphone 18 Ultra"
                    },
                    "price": {
                        "type": "numeric",
                        "example": 119.95
                    },
                    "stock": {
                        "type": "integer",
                        "example": 15
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}